おみくじの処理を別ファイルに分離
